allow edit geometry directly
